Reduce coupling of application with dependency injection
Remove CommandFactory class (logic is handled by AppContext)
Remove DaoFactory class (logic is handled by AppContext)
